Revert "type check classes according to pre order traversal"
This reverts commit e9bbdec05eee1cf4cfdfdcf952ed9f4d1c12786b.
and commit 1a46cc6f4cfd2eadc713cce071d8b85f88ac557a

A better approach might be to make a pass before type checking
to gather attributes and methods in the symbol table.
This way apart from solving the problem of inherited atts and methods
, if a class is used before being defined
for e.g an object of that class is made and its method is accessed
it would be in the symbol table.